+++ core/package.json
@@ -6,26 +6,18 @@
 		"docs": "rimraf ../docs/ && npx typedoc --tsconfig tsconfig.json"
 	},
 	"dependencies": {
+		"@balena/leviathan-test-helpers": "^1.1.1",
 		"ajv": "^6.5.0",
 		"ajv-semver": "^1.0.4",
-		"any-promise": "^1.3.0",
-		"balena-image-fs": "^7.0.6",
-		"balena-sdk": "^16.12.1",
 		"bluebird": "^3.5.1",
-		"bluebird-retry": "^0.11.0",
 		"express": "^4.17.1",
 		"express-ws": "^4.0.0",
 		"fs-extra": "^3.0.1",
 		"lodash": "^4.17.10",
 		"md5-file": "^4.0.0",
 		"mz": "^2.7.0",
-		"node-ssh": "^12.0.4",
-		"request": "^2.88.0",
-		"request-promise": "^4.2.4",
-		"ssh-keygen-lite": "^1.3.0",
 		"tap": "^14.10.8",
 		"tar-fs": "^2.0.0",
-		"tmp": "^0.2.1",
 		"ws": "^7.1.2"
 	},
 	"devDependencies": {
